Unpin Flask
So, on Ubuntu 18.04 with Python 3.6.9 (yeah, both old, gotta upgrade)
the Flask I get is 2.0.3. But, there's a security fix that hasn't been
backported. Dependabot kindly supplies a later version with the fix.
But, surprise!, that newer version of Flask isn't compatible with
Python 3.6.9. So, for the time being, unpin.